% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MAGIC_data}
\alias{MAGIC_data}
\title{MAGIC data}
\format{
An object of class \code{data.frame} with 19020 rows and 11 columns.
}
\usage{
MAGIC_data
}
\description{
MAGIC data
}
\keyword{datasets}
