% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbbr_train.R
\name{rbbr_train}
\alias{rbbr_train}
\title{This function trains RBBR as a supervised artificial intelligence model to learn Boolean rules.}
\usage{
rbbr_train(
  data,
  max_feature = NA,
  mode = NA,
  slope = NA,
  weight_threshold = NA,
  balancing = NA,
  num_cores = NA
)
}
\arguments{
\item{data}{The dataset with scaled features within the [0,1] interval. Each row represents a sample and each column represents a feature. The target variable must be in the last column.}

\item{max_feature}{The maximum number of input features allowed in a Boolean rule. The default value is 3.}

\item{mode}{Choose between "1L" for fitting 1-layered models or "2L" for fitting 2-layered models. The default value is "1L".}

\item{slope}{The slope parameter used in the Sigmoid activation function. The default value is 10.}

\item{weight_threshold}{Conjunctions with weights above this threshold in the fitted ridge regression models will be printed as active conjunctions in the output. The default value is 0.}

\item{balancing}{This is for adjusting the distribution of classes or categories within a dataset to ensure that each class is adequately represented. The default value is "True". Set it to "False", if you don't need to perform the data balancing.}

\item{num_cores}{Specify the number of parallel workers (adjust according to your system)}
}
\value{
This function outputs the predicted Boolean rules with the best Bayesian Information Criterion (BIC).
}
\description{
Regression-Based Boolean Rule (RBBR) inference is performed on datasets where the target feature is either binarized or continuous within the [0,1] range.
}
\examples{
\donttest{
# Load dataset
data(example_data)

# Inspect loaded data
head(XOR_data)

data_train   <- XOR_data[1:800, ]
data_test    <- XOR_data[801:1000, ]

# training model
trained_model<- rbbr_train(data_train)

head(trained_model$boolean_rules)

# testing model
data_test_x  <- data_test[ ,1:(ncol(data_test)-1)]
labels       <- data_test[ ,ncol(data_test)]

predicted_labels <- rbbr_predictor(trained_model, data_test_x, num_top_rules = 1, num_cores = NA)
head(predicted_labels) # predicted labels
head(labels) # true labels

}
}
