% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbbr_predictor.R
\name{rbbr_predictor}
\alias{rbbr_predictor}
\title{This function is designed to make predictions on a new dataset by utilizing a trained model.}
\usage{
rbbr_predictor(
  trained_model,
  data_test,
  num_top_rules = NA,
  slope = NA,
  num_cores = NA
)
}
\arguments{
\item{trained_model}{The model that has been trained using the rbbr_train() function on a training dataset.}

\item{data_test}{The new dataset for which we want to predict the target value or label probability. Each sample is represented as a row, and the input features are in columns.}

\item{num_top_rules}{Specifies the number of Boolean rules with the best Bayesian Information Criterion (BIC) scores to be used for prediction. The default value is 1.}

\item{slope}{The slope parameter used in the Sigmoid activation function. The default value is 10.}

\item{num_cores}{Specify the number of parallel workers (adjust according to your system).}
}
\value{
This function returns the predicted target value or label probability for the input dataset. For each sample in the input dataset (represented as rows), rbbr_predictor() provides the corresponding predicted value or label probability.
}
\description{
This function is designed to make predictions on a new dataset by utilizing a trained model.
}
