% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbbr_scaling.R
\name{rbbr_scaling}
\alias{rbbr_scaling}
\title{This function adjusts the scale of input features to fit within the range of [0,1].}
\usage{
rbbr_scaling(data)
}
\arguments{
\item{data}{The original dataset. Each row corresponds to a sample, and each column represents a feature. The target variable (binary label or continuous value) is expected to be in the last column.}
}
\value{
The function returns the scaled features, ensuring they fall within the range of [0,1].
}
\description{
This function adjusts the scale of input features to fit within the range of [0,1].
}
\examples{
# Load dataset
data(example_data)

# Inspect loaded data
head(MAGIC_data)

data_scaled   <- rbbr_scaling(MAGIC_data)
head(data_scaled)
}
